// src/app/models/padron-ruc.ts
export class PadronRuc {
    ruc?: string;
    razonSocial?: string;
    estado?: string;
    condicion?: string;
    ubigeo?: string;
    tipoVia?: string;
    nombreVia?: string;
    codZona?: string;
    tipoZona?: string;
    numero?: string;
    interior?: string;
    lote?: string;
    departamento?: string;
    manzana?: string;
    kilometro?: string;
}